+++ packages/web-components/src/tab-panel/tab-panel.template.ts
@@ -1,0 +1,3 @@
+import { tabPanelTemplate } from '@microsoft/fast-foundation';
+
+export const template = tabPanelTemplate();